pyproject.toml
# Poetry project definition aligned with Stage 0 requirements.
[tool.poetry]
name = "tg-auto-bot"
version = "0.1.0"
description = "Lite Telegram broadcast automation bot"
readme = "README.md"
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
apscheduler = "^3.10.4"
fastapi = "^0.112.0"
feedparser = "^6.0.11"
httpx = "^0.27.0"
loguru = "^0.7.2"
python-telegram-bot = "^21.4"
python-dotenv = "^1.0.1"
sqlmodel = "^0.0.16"
uvicorn = {extras = ["standard"], version = "^0.30.3"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.23.7"
pytest-cov = "^5.0.0"

[tool.pytest.ini_options]
addopts = "-ra --strict-markers"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = ["asyncio: mark test as requiring asyncio event loop"]

[tool.black]
line-length = 88
target-version = ["py311"]

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
